refactor(minglit_kit): split event_repository_queries into feed + checkin mixins
event_repository_queries.dart (758→212 lines) held mixed concerns:
feed queries (searchEvents, getEventsByType) alongside check-in queries
(getEventById, getTicketBalanceStatus, getEntryGroupParticipantCounts).

Split:
- searchEvents, getEventsByType → _EventRepositoryFeedQueries
  (event_repository_feed_queries.dart, already existed)
- getEventById, getTicketBalanceStatus, getEntryGroupParticipantCounts →
  new _EventRepositoryCheckinQueries (event_repository_checkin_queries.dart)
- Delete event_repository_queries.dart (now empty)
- Remove stale _EventRepositoryQueries constraint from _EventRepositoryCommands
  (commands never called any query method — constraint was leftover)

No logic changes; public API of EventRepository unchanged.

Closes #1965

Constraint: Part files inherit imports from event_repository.dart —
  no explicit imports needed in the split files
Confidence: high
Scope-risk: narrow
